cors

+++ src/main/java/com/coppel/api/denuncias/api_denuncias/Repositories/DenunciaRepository.java
@@ -3,7 +3,7 @@ package com.coppel.api.denuncias.api_denuncias.Repositories;
 import com.coppel.api.denuncias.api_denuncias.Entities.Denuncia;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DenunciaRepository extends JpaRepository<Denuncia, Long> {
+public interface DenunciaRepository extends JpaRepository<Denuncia, Integer> {
 
     Denuncia findByFolio(String folio);
